feat: move logic to a separate file
